module: introduce signal_error method in mod_scmi_from_transport_api
In preparation to allow services other than SCMI to use SMT as a
transport interface, defer the handling of the error related to the
size/length of the payload on a SMT channel to a service that subscribes
to the channel. This relieves SMT module from having to respond to a
malformed payload on behalf of SCMI service.

Change-Id: Ie94086045eb2ebc4692044c398f6dcfd3dc3acf2
Signed-off-by: Aditya Angadi <aditya.angadi@arm.com>